requirements.txt
objectrest~=1.1.1
pydantic==1.9.0
pytz~=2021.1
python-dotenv~=0.19.2
packaging~=21.3